+++ source/spritetools_entity.c
@@ -12,7 +12,7 @@
 
 /* Takes 2 strings and compares them */
 /* Returns 1 on match, 0 on difference */
-static int mystrcmp(char *strA, char *strB) /* bc standard strcmp is dumb */
+static u8 mystrcmp(char *strA, char *strB) /* bc standard strcmp is dumb */
 {
   while (*strA == *strB)
   {
@@ -60,7 +60,7 @@ st_entity *ST_EntityCreateEntity(s64 x, s64 y, u8 animCount)
 /* Takes a pointer to a frame */
 void ST_EntityFreeEntity(st_entity *entity)
 {
-  int i;
+  u8 i;
   for (i = 0; i < entity->animationCount; i++)
   {
     ST_AnimationFreeAnimation(entity->animations[i]);
@@ -72,7 +72,7 @@ void ST_EntityFreeEntity(st_entity *entity)
 /* Takes a pointer to an entity, a pointer to an animation, */
 /*   and a name for the animation */
 /* Returns 1 on success and 0 on failure */
-int ST_EntityAddAnimation(st_entity *entity, st_animation *anim, char *name)
+u8 ST_EntityAddAnimation(st_entity *entity, st_animation *anim, char *name)
 {
   if (entity->animationCount < entity->totalAnims)
   {
@@ -168,7 +168,7 @@ void ST_EntitySetColor(st_entity *entity, u8 red, u8 green, u8 blue, u8 alpha)
 /* Sets the direction of an entity */
 /* Takes a pointer to an entity and a direction */
 /* Returns 1 on success and 0 on failure */
-int ST_EntitySetDirection(st_entity *entity, char *dir)
+u8 ST_EntitySetDirection(st_entity *entity, char *dir)
 {
   if (mystrcmp(dir, "east") || mystrcmp(dir, "south east") ||
     mystrcmp(dir, "south") || mystrcmp(dir, "south west") ||
@@ -189,7 +189,7 @@ int ST_EntitySetDirection(st_entity *entity, char *dir)
 /*   2 = west */
 /*   3 = north */
 /* Returns 1 on success and 0 on failure */
-int ST_EntitySetDirectionId(st_entity *entity, u8 dir)
+u8 ST_EntitySetDirectionId(st_entity *entity, u8 dir)
 {
   if (dir == 0)
   {
@@ -238,9 +238,9 @@ int ST_EntitySetDirectionId(st_entity *entity, u8 dir)
 /* Sets the current animation of an entity by name */
 /* Takes a pointer to an entity and the name of the animation to set */
 /* Returns 1 on success and 0 on failure */
-int ST_EntitySetAnimationName(st_entity *entity, char *name)
+u8 ST_EntitySetAnimationName(st_entity *entity, char *name)
 {
-  int i;
+  u8 i;
   for (i = 0; i < entity->animationCount; i++)
   {
     if (mystrcmp(entity->names[i], name))
@@ -262,7 +262,7 @@ int ST_EntitySetAnimationName(st_entity *entity, char *name)
 /* Sets the current animation of an entity by id */
 /* Takes a pointer to an entity and the id of the animation to set */
 /* Returns 1 on success and 0 on failure */
-int ST_EntitySetAnimationId(st_entity *entity, u8 id)
+u8 ST_EntitySetAnimationId(st_entity *entity, u8 id)
 {
   if (id < entity->animationCount)
   {
@@ -366,7 +366,7 @@ void ST_EntityModifyDirection(st_entity *entity, s8 dir)
 {
   if (dir > 0)
   {
-    int i;
+    u8 i;
     for (i = 0; i < dir; i++)
     {
       if (mystrcmp(entity->dir, "east"))
@@ -389,7 +389,7 @@ void ST_EntityModifyDirection(st_entity *entity, s8 dir)
   }
   if (dir < 0)
   {
-    int i;
+    u8 i;
     for (i = 0; i > dir; i--)
     {
       if (mystrcmp(entity->dir, "east"))
@@ -428,7 +428,7 @@ void ST_EntityModifyRotationNoWrap(st_entity *entity, double rotation)
     entity->rotation = 0;
 }
 
-void ST_EntityModifyRedNoWrap(st_entity *entity, int red)
+void ST_EntityModifyRedNoWrap(st_entity *entity, u8 red)
 {
   if (red > 0)
   {
@@ -454,7 +454,7 @@ void ST_EntityModifyRedNoWrap(st_entity *entity, int red)
   }
 }
 
-void ST_EntityModifyGreenNoWrap(st_entity *entity, int green)
+void ST_EntityModifyGreenNoWrap(st_entity *entity, u8 green)
 {
   if (green > 0)
   {
@@ -480,7 +480,7 @@ void ST_EntityModifyGreenNoWrap(st_entity *entity, int green)
   }
 }
 
-void ST_EntityModifyBlueNoWrap(st_entity *entity, int blue)
+void ST_EntityModifyBlueNoWrap(st_entity *entity, u8 blue)
 {
   if (blue > 0)
   {
@@ -506,7 +506,7 @@ void ST_EntityModifyBlueNoWrap(st_entity *entity, int blue)
   }
 }
 
-void ST_EntityModifyAlphaNoWrap(st_entity *entity, int alpha)
+void ST_EntityModifyAlphaNoWrap(st_entity *entity, u8 alpha)
 {
   if (alpha > 0)
   {
@@ -533,7 +533,7 @@ void ST_EntityModifyAlphaNoWrap(st_entity *entity, int alpha)
 }
 
 void ST_EntityModifyColorNoWrap(st_entity *entity,
-  int red, int green, int blue, int alpha)
+  u8 red, u8 green, u8 blue, u8 alpha)
 {
   ST_EntityModifyRedNoWrap(entity, red);
   ST_EntityModifyGreenNoWrap(entity, green);
